refactor traits and implement a lock per repo
Adds a lock per repo so that only a single repo request can be processed
at a time. This makes it impossible for webhooks to be run concurrently
when for a single repo. Previously it was possible for webhooks to run
together which sometimes caused issues if multiple events updated the
same PR.